+++ hoardgen.py
@@ -13,7 +13,7 @@ import configs
 general = configs.CFG(configs.GENERAL)
 generators = configs.CFG(configs.HOARD_GENERATORS)
 
-MAX_CR = general.getInt("General", "max_cr")
+MAX_CR = general.getInt("Generators", "max_cr")
 
 # Now load our sub-modules.
 import staticitem # For static, pre-made magic items
@@ -90,7 +90,7 @@ class Hoard:
                 return gen
                 # We can stop looking now.
         logging.error("No valid generator found. Using default.")
-        gen = general.get("General", "default_generator")
+        gen = general.get("Generators", "default_generator")
         return gen
                 
     def getValue(self, target):